Admin dialogs are now Vue files
